Extracted out one component from NodeNavigationHistoryPane and moved pane and the new component inside different folders with semantic names. Extracted out the function printing the the node navigation history to console and moved it to a utilities script file.
