+++ src/tools/utilities.js
@@ -1,0 +1,6 @@
+import { get } from 'svelte/store'
+import { NodeNavigationHistoryStore } from '../stores/nodeNavigationHistory'
+
+export const consoleLogNavigationHistory = () => {
+    console.log(get(NodeNavigationHistoryStore))
+}
